+++ 37_del3/src/bræt/AbstractFelter.java
@@ -9,15 +9,11 @@ public abstract class  AbstractFelter {
 
 	public AbstractFelter(int feltnr){
 		this.feltnr = feltnr;
-		opretFelt();
 	}
 	
-	public String toString(){
-		String result = "Du har landet på " + feltnavn + ", vent til din næste tur.";
-		return result;
-	}
+	public abstract String toString();
 	
-	private abstract void opretFelt();
+	//public abstract void opretFelt();
