+++ 37_del3/src/bræt/Bræt.java
@@ -16,8 +16,8 @@ import gui_fields.GUI_Street;
  *
  */
 
-public class Bræt {
-
+public class Bræt {	
+	
 	/**
 	 * Laver alle felterne til spillepladen
 	 * @return Felterne til spillepladen i et array
@@ -25,6 +25,23 @@ public class Bræt {
 	public GUI_Field[] lavBræt() {
 		
 		GUI_Field[] fields = new GUI_Field[24];
+		
+//		for(int i=0; i<felter.length; i++){
+//			switch (i) {
+//			case 0:
+//				
+//				break;
+//			case 3:
+//			case 4:
+//				ssdfisld
+//				break;
+//			default:
+//				break;
+//			}
+//			felter[i] = new Ejendom(i).opretFeltt();
+//			
+//		}
+		
 		
 		fields[0] = new GUI_Start();
 		fields[0].setTitle("Start");
