+++ 37_del3/src/bræt/Fængsel.java
@@ -1,18 +1,31 @@
 package bræt;
 
 import entity.Spiller;
+import gui_fields.GUI_Field;
+import gui_fields.GUI_Jail;
 
 public class Fængsel extends AbstractFelter {
 
 	private boolean besøg;
 	private final int kaution = -1;
 	private int feltnr;
+	private GUI_Field[] felter;
 
 	//Konstruktør kan laves
 	public Fængsel( int feltnr){
 		super(feltnr);
 	}
 
+//	@Override
+//	public void opretFelt(){
+//		felter[feltnr] = new GUI_Jail();
+//		felter[feltnr].setTitle("Gå i fængsel");
+//		felter[feltnr].setDescription("");
+//		felter[feltnr].setSubText("Gå i fængsel");
+//	}
+	
+	
+	
 	@Override
 	public String toString(){
 		return "Du er blevet taget for at stjæle slik, og skal i fængsel";
@@ -23,7 +36,7 @@ public class Fængsel extends AbstractFelter {
 	}
 	
 	public void betalKaution(){
-		Spiller.opdater(kaution);
+		Spiller.setLikvideMidler(kaution);
 	}
